requirements.txt
Django==1.8.4
Pillow==2.9.0
idna==2.0
pytils==0.3
requests==2.7.0
six==1.9.0
sorl-thumbnail==12.3
